actually fix vs2019, i hope

+++ test/functional/lib/tests.cpp
@@ -4700,7 +4700,7 @@ GenericRxFuzzForwardGro(
             Params.DataTrailer = std::rand() % (MaxDataTrailer + 1);
         }
 
-        std::vector<UCHAR> Payload((UINT32)Params.PayloadLength + Params.DataTrailer);
+        std::vector<UCHAR> Payload((SIZE_T)Params.PayloadLength + Params.DataTrailer);
         std::generate(Payload.begin(), Payload.end(), []{ return (UCHAR)std::rand(); });
 
         std::vector<UCHAR> PacketBuffer(
